Issue #116: Move nltk and spacy to separate file nlp.py.
